requirements.txt
pygame-ce == 2.3.0
websockets == 11.0.3